updates tests
